pyproject.toml
[tool.poetry]
name = "terraform-agent"
version = "1.0.0"
description = "FastAPI + Celery Terraform Agent with Redis, Infracost, GitHub, policy scanners, and Prometheus."
authors = ["Abhinav Mishra <you@example.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.11"

# Web/API
fastapi = "^0.115.0"
uvicorn = { version = "^0.30.0", extras = ["standard"] }
python-dotenv = "^1.0.1"
requests = "^2.32.3"
rich = "^13.7.1"
httpx = "^0.27.2"
python-multipart = "^0.0.9"

# Config & validation (Pydantic v2)
pydantic = "^2.8.0"
pydantic-settings = "^2.2.1"

# Background tasks
celery = { version = "^5.4.0", extras = ["redis"] }
redis = "^5.0.7"
kombu = "^5.3.7"
billiard = "^4.2.0"

# Auth & tokens
PyJWT = "^2.9.0"

# Observability
prometheus-client = "^0.20.0"

# IaC tooling integrations
infracost = "^0.10.31"
checkov = "^3.2.102"

# Git/GitHub
GitPython = "^3.1.43"
PyGithub = "^2.4.0"

# Google / Gemini (keep as used in your code)
google-generativeai = "^0.7.2"
google-api-python-client = "^2.140.0"
google-auth = "^2.33.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"

# Utilities
cachetools = "^5.3.3"
markdownify = "^0.12.1"
tenacity = "^9.0.0"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
isort = "^5.13.2"
pytest = "^8.3.2"
mypy = "^1.11.1"
types-requests = "^2.32.0.20240914"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

# --- Formatting configs you already had ---
[tool.isort]
profile = "black"

[tool.black]
line-length = 100
target-version = ['py311']
